refactor(fuse): extract pure AST helpers into farray.json.Ast (step 1/N)
First step toward getting the JSON decoder out of the core fusion macro: lift
the tiny, stateless q.reflect AST walks — unwrap, productFields, decomposeLambda
/2, collectPaths, substParam/substPaths, projPath/fieldAccess/isFieldSelect/
isProductCtor, the Path alias — into a shared `farray.json.Ast` object. These
are engine-neutral (no Shape/Ctx/Stage) and JSON-neutral; they're the one place
both the fusion optimizer and the byte decoder agree on how to read a record's
fields off a lambda. Core keeps one-line forwarders so its ~40 call sites are
unchanged. Pure relocation, behaviour-neutral: JSON + parity tests green.
